# docs/api-coverage/openapi-non-prod.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
...
  "paths" : {
...
    "/api/projects/{projectUuid}" : {
      "get" : {
        "operationId" : "getProject",
        "parameters" : [ {
          "in" : "path",
          "name" : "projectUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseGameProjectVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "game-controller" ]
      },
      "put" : {
        "operationId" : "updateProject",
        "parameters" : [ {
          "in" : "path",
          "name" : "projectUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GameProjectRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseGameProjectVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "game-controller" ]
      }
    },
...
    "/api/projects/{projectUuid}/exports/{jobUuid}" : {
      "get" : {
        "operationId" : "get_4",
        "parameters" : [ {
          "in" : "path",
          "name" : "projectUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "jobUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePrototypeExportJobVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "prototype-export-controller" ]
      }
    },
    "/api/projects/{projectUuid}/exports/{jobUuid}/download" : {
      "get" : {
        "operationId" : "download",
        "parameters" : [ {
          "in" : "path",
          "name" : "projectUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "jobUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "format" : "byte",
                  "type" : "string"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "prototype-export-controller" ]
      }
    },
    "/api/projects/{projectUuid}/exports/{jobUuid}/retry" : {
      "post" : {
        "operationId" : "retry_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "projectUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "jobUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePrototypeExportJobVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "prototype-export-controller" ]
      }
    },
...
    "/api/projects/{projectUuid}/playtest-sessions/{sessionUuid}" : {
      "get" : {
        "operationId" : "get_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "projectUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "sessionUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePlaytestSessionVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "playtest-telemetry-controller" ]
      }
    },
    "/api/projects/{projectUuid}/playtest-sessions/{sessionUuid}/events" : {
      "post" : {
        "operationId" : "ingest",
        "parameters" : [ {
          "in" : "path",
          "name" : "projectUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "sessionUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TelemetryBatchRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseTelemetryBatchVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "playtest-telemetry-controller" ]
      }
    },
...
    "/api/projects/{projectUuid}/prototype-versions/compare" : {
      "get" : {
        "operationId" : "compare",
        "parameters" : [ {
          "in" : "path",
          "name" : "projectUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "left",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "right",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePrototypeVersionComparisonVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "prototype-version-controller" ]
      }
    },
...
    "/api/projects/{projectUuid}/prototype-versions/{versionUuid}/approval" : {
      "post" : {
        "operationId" : "decide",
        "parameters" : [ {
          "in" : "path",
          "name" : "projectUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "versionUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PrototypeApprovalRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePrototypeApprovalVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "prototype-approval-controller" ]
      }
    },
    "/api/projects/{projectUuid}/prototype-versions/{versionUuid}/balance-suggestions" : {
      "post" : {
        "operationId" : "suggest",
        "parameters" : [ {
          "in" : "path",
          "name" : "projectUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "versionUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseBalanceSuggestionVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "playtest-telemetry-controller" ]
      }
    },
    "/api/projects/{projectUuid}/prototype-versions/{versionUuid}/exports" : {
      "post" : {
        "operationId" : "create_3",
        "parameters" : [ {
          "in" : "path",
          "name" : "projectUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "versionUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePrototypeExportJobVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "prototype-export-controller" ]
      }
    },
...
    "/api/projects/{projectUuid}/prototype-versions/{versionUuid}/playtest-sessions" : {
      "post" : {
        "operationId" : "create_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "projectUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "versionUuid",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreatePlaytestSessionRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePlaytestSessionVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "playtest-telemetry-controller" ]
      }
    },
...
    "/api/v1/analytics/prompt-versions" : {
      "get" : {
        "operationId" : "list_4",
        "parameters" : [ {
          "in" : "query",
          "name" : "projectId",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "agentType",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "from",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "includeMock",
          "required" : false,
          "schema" : {
            "default" : false,
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListPromptVersionMetricVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "prompt-version-analytics-controller" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ApiResponseBalanceSuggestionVO" : {
        "properties" : {
          "code" : {
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/BalanceSuggestionVO"
          },
          "message" : {
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseGameProjectVO" : {
        "properties" : {
          "code" : {
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/GameProjectVO"
          },
          "message" : {
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseListPromptVersionMetricVO" : {
        "properties" : {
          "code" : {
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "items" : {
              "$ref" : "#/components/schemas/PromptVersionMetricVO"
            },
            "type" : "array"
          },
          "message" : {
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ApiResponsePlaytestSessionVO" : {
        "properties" : {
          "code" : {
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/PlaytestSessionVO"
          },
          "message" : {
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ApiResponsePrototypeApprovalVO" : {
        "properties" : {
          "code" : {
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/PrototypeApprovalVO"
          },
          "message" : {
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ApiResponsePrototypeExportJobVO" : {
        "properties" : {
          "code" : {
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/PrototypeExportJobVO"
          },
          "message" : {
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ApiResponsePrototypeVersionComparisonVO" : {
        "properties" : {
          "code" : {
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/PrototypeVersionComparisonVO"
          },
          "message" : {
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseTelemetryBatchVO" : {
        "properties" : {
          "code" : {
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/TelemetryBatchVO"
          },
          "message" : {
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "BalanceSuggestionVO" : {
        "properties" : {
          "artifactUuid" : {
            "type" : "string"
          },
          "configDigest" : {
            "type" : "string"
          },
          "prototypeVersionUuid" : {
            "type" : "string"
          },
          "recommendation" : {
            "type" : "string"
          },
          "reused" : {
            "type" : "boolean"
          },
          "sampleSize" : {
            "format" : "int32",
            "type" : "integer"
          },
          "snapshotAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "source" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "CreatePlaytestSessionRequest" : { },
...
      "GameProjectRequest" : {
        "properties" : {
          "description" : {
            "minLength" : 1,
            "type" : "string"
          },
          "gameType" : {
            "minLength" : 1,
            "type" : "string"
          },
          "name" : {
            "minLength" : 1,
            "type" : "string"
          },
          "targetPlatform" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "description", "gameType", "name", "targetPlatform" ],
        "type" : "object"
      },
      "GameProjectVO" : {
        "properties" : {
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "gameType" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          },
          "projectUuid" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "targetPlatform" : {
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "userId" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "ParameterDifference" : {
        "properties" : {
          "changed" : {
            "type" : "boolean"
          },
          "key" : {
            "type" : "string"
          },
          "leftValue" : { },
          "rightValue" : { }
        },
        "type" : "object"
      },
...
      "PlaytestSessionVO" : {
        "properties" : {
          "collectedCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "durationMs" : {
            "format" : "int32",
            "type" : "integer"
          },
          "endedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "eventCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "failureReason" : {
            "type" : "string"
          },
          "hitCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "outcome" : {
            "type" : "string"
          },
          "prototypeVersionUuid" : {
            "type" : "string"
          },
          "restartCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "score" : {
            "format" : "int32",
            "type" : "integer"
          },
          "sessionUuid" : {
            "type" : "string"
          },
          "startedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "PromptVersionMetricVO" : {
        "properties" : {
          "callCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "costMissingCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "estimatedCost" : {
            "type" : "number"
          },
          "inputTokens" : {
            "format" : "int64",
            "type" : "integer"
          },
          "insufficientSample" : {
            "type" : "boolean"
          },
          "latencyMissingCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "meanLatencyMs" : {
            "format" : "double",
            "type" : "number"
          },
          "mockSampleCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "outputTokens" : {
            "format" : "int64",
            "type" : "integer"
          },
          "p50LatencyMs" : {
            "format" : "int64",
            "type" : "integer"
          },
          "p95LatencyMs" : {
            "format" : "int64",
            "type" : "integer"
          },
          "promptVersionId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "realSampleCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "successRate" : {
            "format" : "double",
            "type" : "number"
          },
          "unknownMockCount" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "PrototypeApprovalRequest" : {
        "properties" : {
          "decision" : {
            "minLength" : 1,
            "pattern" : "APPROVED|REJECTED",
            "type" : "string"
          },
          "reason" : {
            "maxLength" : 500,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "decision", "reason" ],
        "type" : "object"
      },
      "PrototypeApprovalVO" : {
        "properties" : {
          "actorType" : {
            "type" : "string"
          },
          "actorUserId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "approvalUuid" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "decision" : {
            "type" : "string"
          },
          "directorRunUuid" : {
            "type" : "string"
          },
          "prototypeVersionUuid" : {
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          },
          "reused" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
      "PrototypeExportJobVO" : {
        "properties" : {
          "attemptCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "completedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "jobUuid" : {
            "type" : "string"
          },
          "packageDigest" : {
            "type" : "string"
          },
          "packageName" : {
            "type" : "string"
          },
          "packageSize" : {
            "format" : "int64",
            "type" : "integer"
          },
          "prototypeVersionUuid" : {
            "type" : "string"
          },
          "reused" : {
            "type" : "boolean"
          },
          "status" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "PrototypeVersionComparisonVO" : {
        "properties" : {
          "differences" : {
            "items" : {
              "$ref" : "#/components/schemas/ParameterDifference"
            },
            "type" : "array"
          },
          "left" : {
            "$ref" : "#/components/schemas/PrototypeVersionVO"
          },
          "right" : {
            "$ref" : "#/components/schemas/PrototypeVersionVO"
          }
        },
        "type" : "object"
      },
...
      "TelemetryBatchRequest" : {
        "properties" : {
          "batchUuid" : {
            "minLength" : 1,
            "pattern" : "^[0-9a-fA-F-]{36}$",
            "type" : "string"
          },
          "events" : {
            "items" : {
              "$ref" : "#/components/schemas/TelemetryEventRequest"
            },
            "maxItems" : 50,
            "minItems" : 0,
            "type" : "array"
          }
        },
        "required" : [ "batchUuid", "events" ],
        "type" : "object"
      },
      "TelemetryBatchVO" : {
        "properties" : {
          "acceptedCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "batchUuid" : {
            "type" : "string"
          },
          "reused" : {
            "type" : "boolean"
          },
          "session" : {
            "$ref" : "#/components/schemas/PlaytestSessionVO"
          }
        },
        "type" : "object"
      },
      "TelemetryEventRequest" : {
        "properties" : {
          "clientElapsedMs" : {
            "format" : "int32",
            "maximum" : 1800000,
            "minimum" : 0,
            "type" : "integer"
          },
          "eventUuid" : {
            "minLength" : 1,
            "pattern" : "^[0-9a-fA-F-]{36}$",
            "type" : "string"
          },
          "payload" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "sequence" : {
            "format" : "int32",
            "maximum" : 1000,
            "minimum" : 1,
            "type" : "integer"
          },
          "type" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "clientElapsedMs", "eventUuid", "payload", "sequence", "type" ],
        "type" : "object"
      },
...
    }
  }
}